fix(keyboard): don't set layout when setup first starts
Removes a redundant/duplicate call to setLayout() that was causing
the initial layout to be set when the index changed, which happens on
startup (to select the initial layout).

This setLayout() call is unnecessary because the layout is already set
by the ItemDelegate's onClicked handler when the user selects a layout
(which also works for keyboard navigation).

Resolves https://invent.kde.org/plasma/plasma-setup/-/issues/61